Policies premium

+++ ainsurance/policies.php
@@ -7,8 +7,6 @@
  */
 
 
-
-
 include("../include/main.php");
 include("../include/tables.php");
 
@@ -39,6 +37,7 @@ $table->generate_data();
                         <th scope="col"><?php $table->display_order_links('Number', 'inapol_policy_number'); ?></th>
                         <th scope="col"><?php $table->display_order_links('Customer', 'cst_name'); ?></th>
                         <th scope="col"><?php $table->display_order_links('Customer ID', 'cst_identity_card'); ?></th>
+                        <th scope="col"><?php $table->display_order_links('Status', 'inapol_status'); ?></th>
                         <th scope="col">
                             <a href="policy_modify.php">
                                 <i class="fas fa-plus-circle"></i>
@@ -50,18 +49,28 @@ $table->generate_data();
                     <?php
                     while ($row = $table->fetch_data()) {
                         ?>
-                        <tr onclick="editLine(<?php echo $row["inapol_policy_ID"];?>);">
+                        <tr onclick="editLine(<?php echo $row["inapol_policy_ID"]; ?>);">
                             <th scope="row"><?php echo $row["inapol_policy_ID"]; ?></th>
                             <td><?php echo $row["inapol_policy_number"]; ?></td>
                             <td><?php echo $row["cst_name"]; ?></td>
                             <td><?php echo $row["cst_identity_card"]; ?></td>
+                            <td><?php echo $row["inapol_status"]; ?></td>
                             <td>
-                                <a href="policy_modify.php?lid=<?php echo $row["inapol_policy_ID"]; ?>"><i
-                                        class="fas fa-edit"></i></a>&nbsp
-                                <a href="policy_delete.php?lid=<?php echo $row["inapol_policy_ID"]; ?>"
-                                   onclick="ignoreEdit = true;
-                               return confirm('Are you sure you want to delete this policy?');"><i
-                                        class="fas fa-minus-circle"></i></a>
+                                <?php if ($row['inapol_status'] == 'Outstanding') { ?>
+                                    <a href="policy_modify.php?lid=<?php echo $row["inapol_policy_ID"]; ?>"><i class="fas fa-edit"></i></a>&nbsp
+                                    <a href="policy_delete.php?lid=<?php echo $row["inapol_policy_ID"]; ?>"
+                                       onclick="ignoreEdit = true; return confirm('Are you sure you want to delete this policy?');"><i class="fas fa-minus-circle"></i></a>&nbsp
+                                    <a href="policy_change_status.php?lid=<?php echo $row["inapol_policy_ID"]; ?>"><i
+                                                class="fas fa-lock"></i></a>&nbsp
+                                <?php }
+                                if ($row['ínapol_status'] == 'Active') { ?>
+                                    <a href="policy_modify.php?lid=<?php echo $row["inapol_policy_ID"]; ?>"><i
+                                                class="fas fa-eye"></i></a>&nbsp
+                                <?php }
+                                if ($row['ínapol_status'] == 'Cancelled' || $row['inapol_status'] == 'Deleted') { ?>
+                                    <a href="policy_modify.php?lid=<?php echo $row["inapol_policy_ID"]; ?>"><i
+                                                class="fas fa-eye"></i></a>&nbsp
+                                <?php } ?>
                             </td>
                         </tr>
                         <?php
@@ -76,7 +85,8 @@ $table->generate_data();
 </div>
 <script>
     var ignoreEdit = false;
-    function editLine(id){
+
+    function editLine(id) {
         if (ignoreEdit === false) {
             window.location.assign('policy_modify.php?lid=' + id);
         }
